requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
charset-normalizer==3.0.1
einops==0.6.0
filelock==3.9.0
frozenlist==1.3.3
fsspec==2023.1.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
huggingface-hub==0.12.1
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
joblib==1.2.0
lazy-loader==0.1
lightning-utilities==0.7.1
lpips==0.1.4
markdown==3.4.1
markupsafe==2.1.2
more-itertools==9.1.0
multidict==6.0.4
networkx==3.0
numpy==1.24.2
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.0
pillow==9.4.0
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytorch-lightning==1.9.3
pywavelets==1.4.1
pyyaml==6.0
qudida==0.0.4
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.1
scipy==1.9.1
six==1.16.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2023.2.27
tokenizers==0.13.2
torch==1.12.1
torchaudio==0.12.1
torchmetrics==0.11.1
torchsummary==1.5.1
torchvision==0.13.1
tqdm==4.64.1
transformers==4.26.1
typing-extensions==4.5.0
urllib3==1.26.14
werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0